fix(ci): Use github.repository for GH_IMAGE to fix release workflow
The release workflow was using `github.actor` which represents the user that triggered the workflow, instead of `github.repository_owner` which represents the owner of the repository. This caused the docker image to be pushed to the wrong location. This commit fixes this issue by using `github.repository` to construct the `GH_IMAGE` environment variable, ensuring the image is pushed to the correct repository.