feat: Add AUR support (#10)
* feat: Add AUR support for describe-commit and describe-commit-bin

Adds support for building and publishing to the Arch User Repository (AUR)
with two packages: `describe-commit` (build from source) and
`describe-commit-bin` (prebuilt binary).

* docs(README): Add instructions for installing/uninstalling AUR package

Added instructions for installing and uninstalling the AUR package
using `pamac` and `pacman` respectively.

* fix(ci): Use github.repository for GH_IMAGE to fix release workflow

The release workflow was using `github.actor` which represents the user that triggered the workflow, instead of `github.repository_owner` which represents the owner of the repository. This caused the docker image to be pushed to the wrong location. This commit fixes this issue by using `github.repository` to construct the `GH_IMAGE` environment variable, ensuring the image is pushed to the correct repository.